requirements.txt
# Binance Futures Testnet – Trading Bot
# ── Runtime dependencies ──────────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.6.0
fastapi[all]>=0.110.0
sqlmodel>=0.0.16
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.19.0

# ── Development / testing ─────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.12.0
